LPS-147889 Refactor of some upgrade processes for them to run safely more than once

+++ modules/apps/commerce/commerce-inventory-service/src/main/java/com/liferay/commerce/inventory/internal/upgrade/v2_0_0/CommerceInventoryAuditUpgradeProcess.java
@@ -14,6 +14,7 @@
 
 package com.liferay.commerce.inventory.internal.upgrade.v2_0_0;
 
+import com.liferay.commerce.inventory.internal.upgrade.v2_0_0.util.CommerceInventoryAuditTable;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
@@ -28,9 +29,11 @@ public class CommerceInventoryAuditUpgradeProcess extends UpgradeProcess {
 		_addColumn("CIAudit", "logType", "VARCHAR(75)");
 		_addColumn("CIAudit", "logTypeSettings", "TEXT");
 
-		runSQL("delete from CIAudit");
+		if (hasColumn(CommerceInventoryAuditTable.TABLE_NAME, "description")) {
+			runSQL("delete from CIAudit");
 
-		alterTableDropColumn("CIAudit", "description");
+			alterTableDropColumn("CIAudit", "description");
+		}
 	}
 
 	private void _addColumn(
