+++ modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/upgrade/v1_5_0/CProductExternalReferenceCodeUpgradeProcess.java
@@ -16,6 +16,7 @@ package com.liferay.commerce.product.internal.upgrade.v1_5_0;
 
 import com.liferay.commerce.product.internal.upgrade.base.BaseCommerceProductServiceUpgradeProcess;
 import com.liferay.commerce.product.model.impl.CPDefinitionImpl;
+import com.liferay.commerce.product.model.impl.CProductImpl;
 import com.liferay.portal.kernel.util.StringUtil;
 
 /**
@@ -28,17 +29,19 @@ public class CProductExternalReferenceCodeUpgradeProcess
 	protected void doUpgrade() throws Exception {
 		addColumn("CProduct", "externalReferenceCode", "VARCHAR(75)");
 
-		Class<CProductExternalReferenceCodeUpgradeProcess> clazz =
-			CProductExternalReferenceCodeUpgradeProcess.class;
+		if (hasColumn(CProductImpl.TABLE_NAME, "externalReferenceCode")) {
+			Class<CProductExternalReferenceCodeUpgradeProcess> clazz =
+				CProductExternalReferenceCodeUpgradeProcess.class;
 
-		String template = StringUtil.read(
-			clazz.getResourceAsStream(
-				"dependencies" +
-					"/CProductExternalReferenceCodeUpgradeProcess.sql"));
+			String template = StringUtil.read(
+				clazz.getResourceAsStream(
+					"dependencies" +
+						"/CProductExternalReferenceCodeUpgradeProcess.sql"));
 
-		runSQLTemplateString(template, false);
+			runSQLTemplateString(template, false);
 
-		dropColumn(CPDefinitionImpl.TABLE_NAME, "externalReferenceCode");
+			dropColumn(CPDefinitionImpl.TABLE_NAME, "externalReferenceCode");
+		}
 	}
 
 }
