+++ modules/apps/commerce/commerce-service/src/main/java/com/liferay/commerce/internal/upgrade/v2_2_0/CommerceOrderUpgradeProcess.java
@@ -49,9 +49,11 @@ public class CommerceOrderUpgradeProcess
 	protected void doUpgrade() throws Exception {
 		addColumn("CommerceOrder", "commerceAccountId", "LONG");
 
-		runSQL("update CommerceOrder set groupId = siteGroupId");
+		if (hasColumn(CommerceOrderModelImpl.TABLE_NAME, "siteGroupId")) {
+			runSQL("update CommerceOrder set groupId = siteGroupId");
 
-		dropColumn(CommerceOrderModelImpl.TABLE_NAME, "siteGroupId");
+			dropColumn(CommerceOrderModelImpl.TABLE_NAME, "siteGroupId");
+		}
 
 		if (!hasColumn(
 				CommerceOrderModelImpl.TABLE_NAME, "orderOrganizationId") ||
